requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-dotenv==1.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.5
